refactor: Reordenamiento de archivos en Scripts
